requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Werkzeug==2.3.7
openai==0.28.1
anthropic==0.3.11
requests==2.31.0
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.4
